refactor: rm constants
